--- a/src/openms/source/CHEMISTRY/TheoreticalSpectrumGenerator.cpp
+++ b/src/openms/source/CHEMISTRY/TheoreticalSpectrumGenerator.cpp
@@ -125,36 +125,27 @@ namespace OpenMS
 
   void TheoreticalSpectrumGenerator::getSpectrum(RichPeakSpectrum & spec, const AASequence & peptide, Int charge) const
   {
-    bool add_b_ions(param_.getValue("add_b_ions").toBool());
-    bool add_y_ions(param_.getValue("add_y_ions").toBool());
-    bool add_a_ions(param_.getValue("add_a_ions").toBool());
-    bool add_c_ions(param_.getValue("add_c_ions").toBool());
-    bool add_x_ions(param_.getValue("add_x_ions").toBool());
-    bool add_z_ions(param_.getValue("add_z_ions").toBool());
-
     for (Int z = 1; z <= charge; ++z)
     {
-      if (add_b_ions)
+      if (add_b_ions_)
         addPeaks(spec, peptide, Residue::BIon, z);
-      if (add_y_ions)
+      if (add_y_ions_)
         addPeaks(spec, peptide, Residue::YIon, z);
-      if (add_a_ions)
+      if (add_a_ions_)
         addPeaks(spec, peptide, Residue::AIon, z);
-      if (add_c_ions)
+      if (add_c_ions_)
         addPeaks(spec, peptide, Residue::CIon, z);
-      if (add_x_ions)
+      if (add_x_ions_)
         addPeaks(spec, peptide, Residue::XIon, z);
-      if (add_z_ions)
+      if (add_z_ions_)
         addPeaks(spec, peptide, Residue::ZIon, z);
     }
 
-    bool add_precursor_peaks(param_.getValue("add_precursor_peaks").toBool());
     if (add_precursor_peaks)
     {
       addPrecursorPeaks(spec, peptide, charge);
     }
 
-    bool add_abundant_immonium_ions(param_.getValue("add_abundant_immonium_ions").toBool());
     if (add_abundant_immonium_ions)
     {
       addAbundantImmoniumIons(spec);
@@ -165,8 +156,6 @@ namespace OpenMS
 
   void TheoreticalSpectrumGenerator::addAbundantImmoniumIons(RichPeakSpectrum & spec) const
   {
-    bool add_metainfo(param_.getValue("add_metainfo").toBool());
-
     RichPeak1D p;
 
     // just in case someone wants the ion names;
@@ -175,7 +164,7 @@ namespace OpenMS
     // Histidin immonium ion
     p.setMZ(110.0718);
     p.setIntensity(1.0);
-    if (add_metainfo)
+    if (add_metainfo_)
     {
       String name("iH");
       p.setMetaValue("IonName", name);
@@ -185,7 +174,7 @@ namespace OpenMS
     // Phenylalanin immonium ion
     p.setMZ(120.0813);
     p.setIntensity(1.0);
-    if (add_metainfo)
+    if (add_metainfo_)
     {
       String name("iF");
       p.setMetaValue("IonName", name);
@@ -195,7 +184,7 @@ namespace OpenMS
     // Tyrosine immonium ion
     p.setMZ(136.0762);
     p.setIntensity(1.0);
-    if (add_metainfo)
+    if (add_metainfo_)
     {
       String name("iY");
       p.setMetaValue("IonName", name);
@@ -205,7 +194,7 @@ namespace OpenMS
     // Iso/Leucin immonium ion (same mass for immonium ion)
     p.setMZ(86.09698);
     p.setIntensity(1.0);
-    if (add_metainfo)
+    if (add_metainfo_)
     {
       String name("iL/I");
       p.setMetaValue("IonName", name);
@@ -215,7 +204,7 @@ namespace OpenMS
     // Tryptophan immonium ion
     p.setMZ(159.0922);
     p.setIntensity(1.0);
-    if (add_metainfo)
+    if (add_metainfo_)
     {
       String name("iW");
       p.setMetaValue("IonName", name);
@@ -236,7 +225,6 @@ namespace OpenMS
     Map<double, String> names;
     AASequence ion;
     double intensity(0);
-    bool add_first_prefix_ion(param_.getValue("add_first_prefix_ion").toBool());
 
     // Generate the ion peaks:
     // Does not generate peaks of full peptide (therefore "<").
@@ -247,7 +235,7 @@ namespace OpenMS
     case Residue::AIon:
     {
       Size i = 1;
-      if (!add_first_prefix_ion)
+      if (!add_first_prefix_ion_)
       {
         i = 2;
       }
@@ -258,14 +246,14 @@ namespace OpenMS
         ions[pos] = ion;
         names[pos] = "a" + String(i) + String(charge, '+');
       }
-      intensity = (double)param_.getValue("a_intensity");
+      intensity = a_intensity_; 
       break;
     }
 
     case Residue::BIon:
     {
       Size i = 1;
-      if (!add_first_prefix_ion)
+      if (!add_first_prefix_ion_)
       {
         i = 2;
       }
@@ -276,14 +264,14 @@ namespace OpenMS
         ions[pos] = ion;
         names[pos] = "b" + String(i) + String(charge, '+');
       }
-      intensity = (double)param_.getValue("b_intensity");
+      intensity = b_intensity_;
       break;
     }
 
     case Residue::CIon:
     {
       Size i = 1;
-      if (!add_first_prefix_ion)
+      if (!add_first_prefix_ion_)
       {
         i = 2;
       }
@@ -294,7 +282,7 @@ namespace OpenMS
         ions[pos] = ion;
         names[pos] = "c" + String(i) + String(charge, '+');
       }
-      intensity = (double)param_.getValue("c_intensity");
+      intensity = c_intensity_;
       break;
     }
 
@@ -307,7 +295,7 @@ namespace OpenMS
         ions[pos] = ion;
         names[pos] = "x" + String(i) + String(charge, '+');
       }
-      intensity = (double)param_.getValue("x_intensity");
+      intensity = x_intensity_;
       break;
     }
 
@@ -320,7 +308,7 @@ namespace OpenMS
         ions[pos] = ion;
         names[pos] = "y" + String(i) + String(charge, '+');
       }
-      intensity = (double)param_.getValue("y_intensity");
+      intensity = y_intensity_;
       break;
     }
 
@@ -333,7 +321,7 @@ namespace OpenMS
         ions[pos] = ion;
         names[pos] = "z" + String(i) + String(charge, '+');
       }
-      intensity = (double)param_.getValue("z_intensity");
+      intensity = z_intensity_;
       break;
     }
 
@@ -341,28 +329,21 @@ namespace OpenMS
       cerr << "Cannot create peaks of that ion type" << endl;
     }
 
-    // get the params
-    bool add_losses(param_.getValue("add_losses").toBool());
-    bool add_metainfo(param_.getValue("add_metainfo").toBool());
-    bool add_isotopes(param_.getValue("add_isotopes").toBool());
-    Int max_isotope((Int)param_.getValue("max_isotope"));
-    double rel_loss_intensity((double)param_.getValue("relative_loss_intensity"));
-
     RichPeak1D p;
     for (Map<double, AASequence>::ConstIterator cit = ions.begin(); cit != ions.end(); ++cit)
     {
       ion = cit->second;
       double pos = cit->first;
       String ion_name = names[pos];
-      if (add_isotopes)
+      if (add_isotopes_)
       {
-        IsotopeDistribution dist = ion.getFormula(res_type, charge).getIsotopeDistribution(max_isotope);
+        IsotopeDistribution dist = ion.getFormula(res_type, charge).getIsotopeDistribution(max_isotope_);
         UInt j(0);
         for (IsotopeDistribution::ConstIterator it = dist.begin(); it != dist.end(); ++it, ++j)
         {
           p.setMZ((double)(pos + (double)j * Constants::NEUTRON_MASS_U) / (double)charge);
           p.setIntensity(intensity * it->second);
-          if (add_metainfo && j == 0)
+          if (add_metainfo_ && j == 0)
           {
             p.setMetaValue("IonName", ion_name);
           }
@@ -373,14 +354,14 @@ namespace OpenMS
       {
         p.setMZ(pos);
         p.setIntensity(intensity);
-        if (add_metainfo)
+        if (add_metainfo_)
         {
           p.setMetaValue("IonName", ion_name);
         }
         spectrum.push_back(p);
       }
 
-      if (add_losses)
+      if (add_losses_)
       {
         set<String> losses;
         for (AASequence::ConstIterator it = cit->second.begin(); it != cit->second.end(); ++it)
@@ -396,9 +377,9 @@ namespace OpenMS
         }
 
 
-        if (!add_isotopes)
+        if (!add_isotopes_)
         {
-          p.setIntensity(intensity * rel_loss_intensity);
+          p.setIntensity(intensity * rel_loss_intensity_);
         }
 
         for (set<String>::const_iterator it = losses.begin(); it != losses.end(); ++it)
@@ -422,15 +403,15 @@ namespace OpenMS
           double loss_pos = loss_ion.getMonoWeight() / (double)charge;
           String loss_name = *it;
 
-          if (add_isotopes)
+          if (add_isotopes_)
           {
-            IsotopeDistribution dist = loss_ion.getIsotopeDistribution(max_isotope);
+            IsotopeDistribution dist = loss_ion.getIsotopeDistribution(max_isotope_);
             UInt j(0);
             for (IsotopeDistribution::ConstIterator iso = dist.begin(); iso != dist.end(); ++iso)
             {
               p.setMZ((double)(loss_pos + j) / (double)charge);
-              p.setIntensity(intensity * rel_loss_intensity * iso->second);
-              if (add_metainfo && j == 0)
+              p.setIntensity(intensity * rel_loss_intensity_ * iso->second);
+              if (add_metainfo_ && j == 0)
               {
                 p.setMetaValue("IonName", ion_name + "-" + loss_name);
               }
@@ -440,7 +421,7 @@ namespace OpenMS
           else
           {
             p.setMZ(loss_pos);
-            if (add_metainfo)
+            if (add_metainfo_)
             {
               p.setMetaValue("IonName", ion_name + "-" + loss_name);
             }
@@ -450,7 +431,7 @@ namespace OpenMS
       }
     }
 
-    if (add_metainfo)
+    if (add_metainfo_)
     {
       p.setMetaValue("IonName", String(""));
     }
@@ -462,26 +443,19 @@ namespace OpenMS
 
   void TheoreticalSpectrumGenerator::addPrecursorPeaks(RichPeakSpectrum & spec, const AASequence & peptide, Int charge) const
   {
-    bool add_metainfo(param_.getValue("add_metainfo").toBool());
-    double pre_int((double)param_.getValue("precursor_intensity"));
-    double pre_int_H2O((double)param_.getValue("precursor_H2O_intensity"));
-    double pre_int_NH3((double)param_.getValue("precursor_NH3_intensity"));
-    bool add_isotopes(param_.getValue("add_isotopes").toBool());
-    int max_isotope((int)param_.getValue("max_isotope"));
-
     RichPeak1D p;
 
     // precursor peak
     double mono_pos = peptide.getMonoWeight(Residue::Full, charge) / double(charge);
-    if (add_isotopes)
+    if (add_isotopes_)
     {
-      IsotopeDistribution dist = peptide.getFormula(Residue::Full, charge).getIsotopeDistribution(max_isotope);
+      IsotopeDistribution dist = peptide.getFormula(Residue::Full, charge).getIsotopeDistribution(max_isotope_);
       UInt j(0);
       for (IsotopeDistribution::ConstIterator it = dist.begin(); it != dist.end(); ++it, ++j)
       {
         p.setMZ((double)(mono_pos + j * Constants::NEUTRON_MASS_U) / (double)charge);
-        p.setIntensity(pre_int *  it->second);
-        if (add_metainfo)
+        p.setIntensity(pre_int_ *  it->second);
+        if (add_metainfo_)
         {
           String name("[M+H]+");
           if (charge == 2)
@@ -496,8 +470,8 @@ namespace OpenMS
     else
     {
       p.setMZ(mono_pos);
-      p.setIntensity(pre_int);
-      if (add_metainfo)
+      p.setIntensity(pre_int_);
+      if (add_metainfo_)
       {
         String name("[M+H]+");
         if (charge == 2)
@@ -513,15 +487,15 @@ namespace OpenMS
     //loss of water
     EmpiricalFormula ion = peptide.getFormula(Residue::Full, charge) - EmpiricalFormula("H2O");
     mono_pos = ion.getMonoWeight() / double(charge);
-    if (add_isotopes)
+    if (add_isotopes_)
     {
-      IsotopeDistribution dist = ion.getIsotopeDistribution(max_isotope);
+      IsotopeDistribution dist = ion.getIsotopeDistribution(max_isotope_);
       UInt j(0);
       for (IsotopeDistribution::ConstIterator it = dist.begin(); it != dist.end(); ++it, ++j)
       {
         p.setMZ((double)(mono_pos + j * Constants::NEUTRON_MASS_U) / (double)charge);
-        p.setIntensity(pre_int_H2O *  it->second);
-        if (add_metainfo)
+        p.setIntensity(pre_int_H2O_ *  it->second);
+        if (add_metainfo_)
         {
           String name("[M+H]-H2O+");
           if (charge == 2)
@@ -536,8 +510,8 @@ namespace OpenMS
     else
     {
       p.setMZ(mono_pos);
-      p.setIntensity(pre_int_H2O);
-      if (add_metainfo)
+      p.setIntensity(pre_int_H2O_);
+      if (add_metainfo_)
       {
         String name("[M+H]-H2O+");
         if (charge == 2)
@@ -552,15 +526,15 @@ namespace OpenMS
     //loss of ammonia
     ion = peptide.getFormula(Residue::Full, charge) - EmpiricalFormula("NH3");
     mono_pos = ion.getMonoWeight() / double(charge);
-    if (add_isotopes)
+    if (add_isotopes_)
     {
-      IsotopeDistribution dist = ion.getIsotopeDistribution(max_isotope);
+      IsotopeDistribution dist = ion.getIsotopeDistribution(max_isotope_);
       UInt j(0);
       for (IsotopeDistribution::ConstIterator it = dist.begin(); it != dist.end(); ++it, ++j)
       {
         p.setMZ((double)(mono_pos + j * Constants::NEUTRON_MASS_U) / (double)charge);
-        p.setIntensity(pre_int_NH3 *  it->second);
-        if (add_metainfo)
+        p.setIntensity(pre_int_NH3_ *  it->second);
+        if (add_metainfo_)
         {
           String name("[M+H]-NH3+");
           if (charge == 2)
@@ -575,8 +549,8 @@ namespace OpenMS
     else
     {
       p.setMZ(mono_pos);
-      p.setIntensity(pre_int_NH3);
-      if (add_metainfo)
+      p.setIntensity(pre_int_NH3_);
+      if (add_metainfo_)
       {
         String name("[M+H]-NH3+");
         if (charge == 2)
@@ -589,6 +563,33 @@ namespace OpenMS
     }
 
     spec.sortByPosition();
+  }
+
+  void TheoreticalSpectrumGenerator::updateMembers_()
+  {
+    add_b_ions_ = param_.getValue("add_b_ions").toBool();
+    add_y_ions_ = param_.getValue("add_y_ions").toBool();
+    add_a_ions_ = param_.getValue("add_a_ions").toBool();
+    add_c_ions_ = param_.getValue("add_c_ions").toBool();
+    add_x_ions_ = param_.getValue("add_x_ions").toBool();
+    add_z_ions_ = param_.getValue("add_z_ions").toBool();
+    add_first_prefix_ion_ = param_.getValue("add_first_prefix_ion").toBool();
+    add_losses_ = param_.getValue("add_losses").toBool();
+    add_metainfo_ = param_.getValue("add_metainfo").toBool();
+    add_isotopes_ = param_.getValue("add_isotopes").toBool();
+    add_precursor_peaks = param_.getValue("add_precursor_peaks").toBool();
+    add_abundant_immonium_ions = param_.getValue("add_abundant_immonium_ions").toBool();
+    a_intensity_ = (double)param_.getValue("a_intensity");
+    b_intensity_ = (double)param_.getValue("b_intensity");
+    c_intensity_ = (double)param_.getValue("c_intensity");
+    x_intensity_ = (double)param_.getValue("x_intensity");
+    y_intensity_ = (double)param_.getValue("y_intensity");
+    z_intensity_ = (double)param_.getValue("z_intensity");
+    max_isotope_ = (Int)param_.getValue("max_isotope");
+    rel_loss_intensity_ = (double)param_.getValue("relative_loss_intensity");
+    pre_int_ = (double)param_.getValue("precursor_intensity");
+    pre_int_H2O_ = (double)param_.getValue("precursor_H2O_intensity");
+    pre_int_NH3_ = (double)param_.getValue("precursor_NH3_intensity");
   }
 
 }