--- a/src/tests/class_tests/openms/source/TheoreticalSpectrumGenerator_test.cpp
+++ b/src/tests/class_tests/openms/source/TheoreticalSpectrumGenerator_test.cpp
@@ -108,7 +108,7 @@ START_SECTION(void addPeaks(RichPeakSpectrum& spectrum, const AASequence& peptid
   {
     TEST_REAL_SIMILAR(y_spec2[i].getPosition()[0], (y_result[i]+1.0)/2.0)
   }
-
+/* for quick benchmarking of implementation chances
   for (Size i = 0; i != 1e6; ++i)
   {
     RichPeakSpectrum y_spec, b_spec, a_spec;
@@ -116,7 +116,7 @@ START_SECTION(void addPeaks(RichPeakSpectrum& spectrum, const AASequence& peptid
     ptr->addPeaks(b_spec, peptide, Residue::BIon, 1);
     ptr->addPeaks(a_spec, peptide, Residue::AIon, 1);
   }
-
+*/
 END_SECTION
 
 START_SECTION(void addAbundantImmoniumIons(RichPeakSpectrum& spec))