Creation of 3*10e6 ion ladders for a short peptide without loss ladders and isotope clusters: 
before speedup ca. 50s
introduce updateMembers_: 40s
removed map lockup: 10s
initialize size: 8s