@timosachsenberg 2015-12-21
there are some address sanitizer issues that will be fixed.