@jpfeuffer 2016-02-15
If you need it even faster, have a look at #1406.